Вроде основную часть перекинул. Теперь будем дебажить
