+++ DataImpression/View/CSVOpenMasterView.xaml
@@ -1,0 +1,36 @@
+﻿<Window x:Class="DataImpression.View.CSVOpenMasterView"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:DataImpression.View"
+        mc:Ignorable="d"
+        Title="CSVOpenMasterView" Height="450" Width="800">
+    <Grid>
+        <Grid.RowDefinitions>
+            <RowDefinition Height="40"/>
+            <RowDefinition/>
+            <RowDefinition Height="40"/>
+        </Grid.RowDefinitions>
+        <TextBlock Text="{Binding InputPageTitle}" HorizontalAlignment="Center" VerticalAlignment="Center"/>
+        <local:TimeColumnChoiceView  Grid.Row="1"
+                                         DataContext="{Binding TimeColumnChoiceVM}"
+                                         Visibility="{Binding Visibility, UpdateSourceTrigger=PropertyChanged}"
+                                         />
+        <local:AOIHitColumnsChoiceView Grid.Row="1" 
+                                            DataContext="{Binding AOIHitColumnsChoiceVM}"
+                                            Visibility="{Binding Visibility}"
+                                           />
+        <local:FAOIsInputView x:Name="FAOIsInput" Grid.Row="1"
+                                            DataContext="{Binding FAOIsInputVM}"
+                                            Visibility="{Binding Visibility}"
+                                           />
+        <local:ProcessingTaskView  Grid.Row="1"
+                                            DataContext="{Binding ProcessingTaskVM}"
+                                            Visibility="{Binding Visibility}"
+                                           />
+
+        <Button Grid.Row="2" Content="Далее" HorizontalAlignment="Right"  Width="60" Margin="7"
+                    Command="{Binding NextInputCommand}"/>
+    </Grid>
+</Window>
